requirements.txt
anthropic>=0.40.0
openai>=1.30.0
faiss-cpu>=1.7.4
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0
